Changed TitleBar on DevicesScreen Changed it to look the same as RoutinesScreen etc
