+++ HomeApp/app/src/main/java/com/HomeApp/screens/ChooseActionsScreen.kt
@@ -67,7 +67,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.HomeApp.ui.composables.CustomFAB
-import com.HomeApp.ui.composables.TopTitleBar
+import com.HomeApp.ui.composables.TitleBar
 import com.HomeApp.ui.navigation.ChooseItems
 import com.HomeApp.ui.navigation.ChooseSchedule
 import com.HomeApp.ui.navigation.EditTrigger
@@ -217,7 +217,7 @@ fun ChooseActionsScreen(
 
     Scaffold(
         topBar = {
-            TopTitleBar(
+            TitleBar(
                 title = "Actions",
                 iconLeft = Icons.Rounded.ArrowBack,
                 routeLeftButton = ChooseItems.route,
