+++ HomeApp/app/src/main/java/com/HomeApp/screens/ChooseItemsScreen.kt
@@ -43,7 +43,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.HomeApp.ui.composables.CustomFAB
-import com.HomeApp.ui.composables.TopTitleBar
+import com.HomeApp.ui.composables.TitleBar
 import com.HomeApp.ui.navigation.ChooseActions
 import com.HomeApp.ui.navigation.ChooseType
 import com.HomeApp.ui.navigation.Routines
@@ -155,7 +155,7 @@ fun ChooseItemsScreen(
 
     Scaffold(
         topBar = {
-            TopTitleBar(
+            TitleBar(
                 title = if (isDevices) "Devices" else "Groups",
                 iconLeft = Icons.Rounded.ArrowBack,
                 routeLeftButton = ChooseType.route,
