+++ HomeApp/app/src/main/java/com/HomeApp/screens/ChooseTypeScreen.kt
@@ -27,7 +27,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import com.HomeApp.ui.composables.TopTitleBar
+import com.HomeApp.ui.composables.TitleBar
 import com.HomeApp.ui.navigation.ChooseItems
 import com.HomeApp.ui.navigation.EditTrigger
 import com.HomeApp.ui.navigation.Routines
@@ -46,7 +46,7 @@ fun ChooseTypeScreen(
 
     Scaffold(
         topBar = {
-            TopTitleBar(
+            TitleBar(
                 title = "Choose",
                 iconLeft = Icons.Rounded.ArrowBack,
                 routeLeftButton =
